refactor: modularize commands into dedicated submodules for better structure and maintainability
