+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,11 @@
+// Command modules
+pub mod app_commands;
+pub mod folder_commands;
+pub mod note_commands;
+pub mod vault_commands;
+
+// Re-export all command functions
+pub use app_commands::*;
+pub use folder_commands::*;
+pub use note_commands::*;
+pub use vault_commands::*;
